requirements.txt
# ============================================================================
# CHATTERBOX TTS SERVER - REQUIREMENTS (CPU-ONLY INSTALLATION)
# ============================================================================
# This file ensures compatible PyTorch, ML, and audio processing libraries
# are installed for CPU-only operation. All packages are pinned for stability.

# --- PyTorch CPU-Only Installation Index ---
# Use PyTorch's official CPU-only package index for torch components
--extra-index-url https://download.pytorch.org/whl/cpu

# --- Deep Learning Framework (CPU-Only) ---
# Pinning a compatible set of torch-related libraries is crucial to avoid dependency conflicts.
# Updated to use PyTorch 2.5.1 for compatibility with our Colab-compatible chatterbox fork
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# --- Core TTS Application ---
# Chatterbox TTS engine - Install from chatterbox-v2 fork
chatterbox-tts @ git+https://github.com/devnen/chatterbox-v2.git@master

# --- Web Framework & Server ---
fastapi                         # Modern async web framework
uvicorn[standard]               # ASGI server with performance extras

# --- Scientific Computing & ML Libraries ---
numpy>=1.24.0,<1.26.0           # Fundamental numerical computing
librosa                         # Advanced audio/music analysis
safetensors                     # Safe tensor serialization format
descript-audio-codec            # Audio codec for ML applications

# --- Audio I/O & Processing ---
soundfile                       # Audio file I/O (requires libsndfile system library)
pydub                          # Simple audio manipulation
audiotsm                       # Audio time-scale modification
praat-parselmouth              # Python interface to Praat speech analysis
pyrnnoise                      # RNNoise denoising
auto-editor                    # Automatic silence and stutter removal
openai-whisper                 # Whisper inference
faster-whisper                 # Faster Whisper inference

# --- Web & API Utilities ---
python-multipart               # Form data parsing for FastAPI
requests                       # HTTP client library
Jinja2                        # Template engine
aiofiles                      # Async file operations
hf_transfer                     # Speed up file transfers with the Hugging Face Hub.


# --- Configuration & Data Processing ---
PyYAML                        # YAML configuration file support
watchdog                      # File system event monitoring
unidecode                     # Unicode text transliteration
inflect                       # Natural language inflection
tqdm                          # Progress bars for loops

# --- Background Jobs & Queue ---
redis                         # Redis client
rq                            # Background task queue
python-dotenv                 # Configuration management

# ============================================================================
# INSTALLATION NOTES:
# 
# 1. Ensure your virtual environment is activated before installation
# 2. Run: pip install --upgrade pip
# 3. Run: pip install -r requirements.txt
# 4. For parselmouth issues, the package name is 'praat-parselmouth' on PyPI
#    but you import it as 'import parselmouth' in Python code
# 5. System dependencies may be required:
#    - libsndfile (for soundfile package)
#    - FFmpeg (for some audio processing features)
# ============================================================================
